njila-auth-service/requirements.txt
anyio==4.12.1
asgiref==3.11.1
async-timeout==5.0.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==5.2.12
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
pika==1.3.2
psycopg2-binary==2.9.11
py_eureka_client==0.13.3
PyJWT==2.11.0
python-decouple==3.8
redis==7.3.0
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
